+++ test/test_getGroupDecisionTree.js
@@ -9,34 +9,30 @@ const CONFIGURATION_1_OPERATIONS_1_TO = _.last(CONFIGURATION_1_OPERATIONS_1).tim
 
 describe('client.getGroupDecisionTree(<agentList>, <timestamp>, <configuration>)', function() {
   let client;
-  let agent;
-  const agentId_1 = `get_agent_decision_tree_1_${RUN_ID}`;
-  const agentId_2 = `get_agent_decision_tree_2_${RUN_ID}`;
-  const agentId_list = [agentId_1, agentId_2];
+  const agentId1 = `get_agent_decision_tree_1_${RUN_ID}`;
+  const agentId2 = `get_agent_decision_tree_2_${RUN_ID}`;
+  const notExistingAgentId = `wrong_dude_${RUN_ID}`;
+  const agentIdList = [agentId1, agentId2];
+  const agentIdListWithNonExistingAgent = [agentId1, notExistingAgentId];
+
 
   before(function() {
     client = craftai(CRAFT_CFG);
     expect(client).to.be.ok;
   });
   beforeEach(function() {
-    return client.deleteAgent(agentId_1) // Delete any preexisting agent with this id.
-      .then(() => client.createAgent(CONFIGURATION_1, agentId_1))
-      .then((createdAgent) => {
-        expect(createdAgent).to.be.ok;
-        agent = createdAgent;
-        return client.addAgentContextOperations(agent.id, CONFIGURATION_1_OPERATIONS_1);
-      })
-      .then(() => client.deleteAgent(agentId_2))
-      .then(() => client.createAgent(CONFIGURATION_1, agentId_2))
-      .then((createdAgent) => {
-        expect(createdAgent).to.be.ok;
-        agent = createdAgent;
-        return client.addAgentContextOperations(agent.id, CONFIGURATION_1_OPERATIONS_1);
-      });
+    return Promise.all(agentIdList.map((agentId) => {
+      return client.deleteAgent(agentId) // Delete any preexisting agent with this id.
+        .then(() => client.createAgent(CONFIGURATION_1, agentId))
+        .then((createdAgent) => {
+          expect(createdAgent).to.be.ok;
+          return client.addAgentContextOperations(createdAgent.id, CONFIGURATION_1_OPERATIONS_1);
+        });
+    }));
   });
   
   it('Return a decision tree when called with an agent list', () => {
-    return client.getGroupDecisionTree(agentId_list)
+    return client.getGroupDecisionTree(agentIdList)
       .then((treeJson) => {
         expect(treeJson).to.be.ok;
         const { _version, configuration, trees } = parse(treeJson);
@@ -46,8 +42,17 @@ describe('client.getGroupDecisionTree(<agentList>, <timestamp>, <configuration>)
       });
   });
 
+  it('Throw an error when called with an agent list where an agent does not exist', () => {
+    return client.getGroupDecisionTree(agentIdListWithNonExistingAgent)
+      .then(() => new Error('Should not be reached'))
+      .catch((err) => {
+        expect(err.error).to.be.equal('NotFound');
+        expect(err.status).to.be.equal(404);
+      });
+  });
+
   it('Return a different decision tree when called with a different timestamp', () => {
-    return client.getGroupDecisionTree(agentId_list)
+    return client.getGroupDecisionTree(agentIdList)
       .then((treeJson) => {
         expect(treeJson).to.be.ok;
         const { _version, configuration, trees } = parse(treeJson);
@@ -55,7 +60,7 @@ describe('client.getGroupDecisionTree(<agentList>, <timestamp>, <configuration>)
         expect(_version).to.be.ok;
         expect(configuration).to.be.deep.equal(CONFIGURATION_1);
         const TIME_LAG = 1000;
-        return client.getGroupDecisionTree(agentId_list, { t: CONFIGURATION_1_OPERATIONS_1_TO - TIME_LAG })
+        return client.getGroupDecisionTree(agentIdList, { t: CONFIGURATION_1_OPERATIONS_1_TO - TIME_LAG })
           .then((treeJson2) => {
             expect(treeJson2).to.be.ok;
             const { _version, configuration, trees } = parse(treeJson2);
@@ -70,7 +75,7 @@ describe('client.getGroupDecisionTree(<agentList>, <timestamp>, <configuration>)
   it('Throw an error when a configuration is passed with the wrong context', () => {
     const CONFIGURATION_GROUP = _.cloneDeep(CONFIGURATION_1);
     CONFIGURATION_GROUP.context.lightIntensity.type = 'enum';
-    return client.getGroupDecisionTree(agentId_list, { configuration: CONFIGURATION_GROUP })
+    return client.getGroupDecisionTree(agentIdList, { configuration: CONFIGURATION_GROUP })
       .then(() => Promise.reject(new Error('Should not be reached')))
       .catch((err) => expect(err).to.be.instanceof(errors.CraftAiBadRequestError));
   });
@@ -78,7 +83,7 @@ describe('client.getGroupDecisionTree(<agentList>, <timestamp>, <configuration>)
   it('Success when a good configuration is passed', () => {
     const CONFIGURATION_GROUP = _.cloneDeep(CONFIGURATION_1);
     CONFIGURATION_GROUP.learning_period = 150;
-    return client.getGroupDecisionTree(agentId_list, { configuration: CONFIGURATION_GROUP })
+    return client.getGroupDecisionTree(agentIdList, { configuration: CONFIGURATION_GROUP })
       .then((treeJson) => {
         expect(treeJson).to.be.ok;
         const { _version, configuration, trees } = parse(treeJson);